Set can-persist property
OpenConnect can persist and reconnect when the underlying physical
connection goes away or changes. It does rather depend on NetworkManager
not screwing up the routes and leaving the VPN gateway routed *through*
the VPN itself, and it would be kind of nice to accelerate the reconnect
by sending SIGUSR2 to openconnect to make it do an immediate reconnect
when a physical link comes back up, instead of waiting for Dead Peer
Detection. But basically, it should work.