service: deleage CSD trojan execution to the agent
The service sets --csd-wrapper to a shim program that forwards the agent
to the service. The shim passes the arguments to the service.

The service issues a secrets requests, forwarding the arguments passed
to the shim as hints. The auth helper recognizes those and runs the CSD
trojan instead of asking for actual secrets.

Upon receiving the new "secrets" from the auth helper, the service
responds to the shim which, in turn, prints the result back to the
openconnect.

This makes CSD scripts work with GlobalProtect networks that can request
CSD to be run after the authentication phase.

Based on work by Daniel Lenski <dlenski@gmail.com>