auth-helper: spawn CSD trojan if asked for by the hint
If the auth helper gets a special set of hints ("x-openconnect-csd:*"),
instead of the usual secrets handling, it will compose a CSD command
line from those hints, run the script and pass back the result as if it
were secrets.

This is useful for delegating the CSD trojan execution to the user
session.