service: disable calling the csd-wrapper script from openconnect
NetworkManager's openconnect plugin spawns the openconnect binary not in
the context of the user that activates the profile. That is also due to
the fact that often there is no immediate user around, for example when
the VPN activates automatically.

As such, it is not safe to run code provided by the user.

Note that the plugin changes the user of the openconnect binary to
"nm-openconnect" user. But this is still not safe, because it means that
one user may ptrace/exploit the VPN session activated by another
openconnect user.

Executing the script is basically `sudo -u nm-openconnect $MY_SCRIPT`,
which is not a safe thing to allow to an untrusted user.

This needs a different solution, for example running the script in the
user's context. Which of course then has a different problem, that the
user activating a profile executes a script that may not be configured
by that same user. That might be solvable by only executing the script
if it's owned by the user that activates the profile.

Anyway, whatever the proper solution, for now disable this code to avoid
doing something dangerous. A proper solution needs to be found.

Related: https://gitlab.gnome.org/GNOME/NetworkManager-openconnect/merge_requests/2

Acked-by: Beniamino Galvani <bgalvani@redhat.com>
Acked-by: Lubomir Rintel <lkundrak@v3.sk>