Use return count from openconnect_get_supported_protocols()
From OpenConnect v8.05 to v8.08 this stopped returning the NULL
sentinel at the end of the array, on the flawed assumption that
users would all be using the count that it returns.

That was an error, which will be fixed in the next release. In
the meantime fix up NetworkManager-openconnect to cope with
those versions too.