Allow IP prefixes of 0 for routing rules
A prefix of 0 is used for routing rules that accept all traffic. For
example, a VPN may accept all traffic by default while excluding specific
IPs. Previously, the "accept all traffic by default" rule was rejected as
invalid due to having an IP prefix of 0.

This also requires a similar fix in the main NetworkManager package to work
properly. Versions of NetworkManager without the fix will silently reject
the rule, yielding the same behavior as before this commit.

https://gitlab.gnome.org/GNOME/NetworkManager-openconnect/-/issues/52

https://gitlab.gnome.org/GNOME/NetworkManager-openconnect/-/merge_requests/21