properties: modernize the GtkBuilder file
This reworks the UI file to use GtkGrid instead of a complicated
hierarchy of GtkVBoxes and GtkTables. At the same time it drops
some deprecated properties.

Dropping old widget types and properties will make porting to Gtk4
easier.

Aside from that, the appearance of the labels is made consistent with
what GNOME Control Center and newer versions of nm-connection-editor
use.