Handle default routes in split excludes
We attempted to 'fix' OpenConnect not to send these and to set the netmask
on the interface to 0 instead, but that caused compatibility problems which
we had to work around in commit 84e279cb7 ("src/helper: handle openconnect
8.20 netmask values.")

We want to revert that from the OpenConnect side as it's a regression, so
let's find a better way to achieve the original objective. Scan the split
includes to see if they include a default route. If they *do*, drop it from
the list we pass to NM explicitly, but *don't* set the never-default flag.

That should allow NM to honour the 'Use only for resources on this
connection' setting while still doing the right thing in other cases.